OD-10915 Revert: Removed number extension from Money class to fix groovy relational operators. We don't use Money class as a number.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@49024 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ common/waCommon/src/main/java/ish/math/MoneyDecimalFormatter.java
@@ -15,7 +15,7 @@ public class MoneyDecimalFormatter {
 		xmlDecimal.setGroupingUsed(false);
 		xmlDecimal.setMinimumFractionDigits(scale);
 		xmlDecimal.setMaximumFractionDigits(scale);
-		return xmlDecimal.format(money.toBigDecimal());
+		return xmlDecimal.format(money.doubleValue());
 	}
 
 	public Money stringToValue(String value) {
